Assign Foreigh Key to Document class
